% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrvar.R
\name{lrvar}
\alias{lrvar}
\title{Estimation of the Long-Run, Half Long-Run and One-Sided Long-Run
Covariance Matrix of a Linear process}
\usage{
lrvar(tsrs, krnl = "ba", band = "and")
}
\arguments{
\item{tsrs}{A vector or matrix of values for the time series.}

\item{krnl}{Kernel function used for long-run variance estimation. Either:
\itemize{
\item \code{"tr"}: Truncated
\item \code{"ba"}: Bartlett
\item \code{"pa"}: Parzen
\item \code{"bo"}: Bohman
\item \code{"da"}: Daniell
\item \code{"qs"}: Quadratic Spectral
}

Default is \code{"ba"}.}

\item{band}{Bandwidth specification. Either:
\itemize{
\item Integer in \{1, ..., T\} (explicit bandwidth value)
\item \code{"and"}: Data-dependent rule (Andrews, 1991)
\item \code{"nw"}: Data-dependent rule (Newey & West, 1987)
\item \code{"nwt"}: Rule-of-thumb, i.e., floor(4 * (T / 100)^(2 / 9))
}

Default is \code{"and"}.}
}
\value{
A list containing:
\itemize{
\item \code{bwdh}: The bandwidth computed/specified
\item \code{halfvar}: The half long-run covariance matrix
\item \code{shrtvar}: The contemporaenous covariance matrix
\item \code{longvar}: The long-run covariance matrix
\item \code{deltvar}: The one-sided long-run covariance matrix, i.e.,
\code{halfvar} + \code{shrtvar}
}
}
\description{
Computes long-run, half long-run and one-sided long-run covariance matrix
estimates of a (multivariate) time series
}
\references{
Andrews, D. W. K. (1991).
"Heteroskedasticity and autocorrelation consistent covariance matrix
estimation." \emph{Econometrica}, 59(3), 817-858.

Newey, W. K., and West, K. D. (1987).
"A simple, positive semi-definite, heteroskedasticity and autocorrelation
consistent covariance matrix." \emph{Econometrica}, 55(3), 703-708.
}
\seealso{
bwclc, lrwghts, computelrvec, computelrmat
}
